requirements.txt
requests==2.32.5
pandas==3.0.0
openpyxl==3.1.5
python-dotenv==1.0.1
